test refactor - move the common FeedItemEventListener test helper to the proper event/ package.
