+++ app/src/androidTest/java/de/test/antennapod/util/event/FeedItemEventListener.java
@@ -1,4 +1,4 @@
-package de.test.antennapod.util.feed;
+package de.test.antennapod.util.event;
 
 import android.support.annotation.NonNull;
